Port probing

+++ include/sys/pci.h
@@ -6,7 +6,7 @@
 uint16_t pci_read_word (uint8_t bus, uint8_t slot, uint8_t func, uint8_t offset);
 uint16_t get_vendor_id(uint8_t bus, uint8_t device);
 void check_all_buses();
-uint32_t pci_read_bar (uint8_t bus, uint8_t slot, uint8_t func, uint8_t offset);
+uint64_t pci_read_bar (uint8_t bus, uint8_t slot, uint8_t func, uint8_t offset);
 
 
 #endif
